Adjusted threat assessment logic to handle velocity and time filtering correctly. Updated tests.

+++ src/main/java/com/challenge/ship/coordination/dispatchservice/dto/ShipHistoryPosition.java
@@ -2,7 +2,7 @@ package com.challenge.ship.coordination.dispatchservice.dto;
 
 import com.challenge.ship.coordination.dispatchservice.model.ShipPosition;
 
-public record ShipHistoryPosition(int time, int speed, Position position) {
+public record ShipHistoryPosition(long time, int speed, Position position) {
 
   public ShipHistoryPosition(ShipPosition shipPosition) {
     this(shipPosition.time(), shipPosition.speed(),
