+++ src/main/java/com/challenge/ship/coordination/dispatchservice/dto/ShipPositionRequest.java
@@ -2,6 +2,6 @@ package com.challenge.ship.coordination.dispatchservice.dto;
 
 import com.challenge.ship.coordination.dispatchservice.model.TemporalPosition;
 
-public record ShipPositionRequest(int time, int x, int y) implements TemporalPosition {
+public record ShipPositionRequest(long time, int x, int y) implements TemporalPosition {
 
 }
