+++ src/main/java/com/challenge/ship/coordination/dispatchservice/model/ShipPosition.java
@@ -1,6 +1,6 @@
 package com.challenge.ship.coordination.dispatchservice.model;
 
-public record ShipPosition(int time, int x, int y, int speed, ThreatStatus status)
+public record ShipPosition(long time, int x, int y, int speed, ThreatStatus status)
     implements TemporalPosition {
 
   public ShipPosition(TemporalPosition temporalPosition, int speed, ThreatStatus status) {
