+++ src/main/java/com/challenge/ship/coordination/dispatchservice/model/TemporalPosition.java
@@ -6,6 +6,6 @@ public interface TemporalPosition {
 
   int y();
 
-  int time();
+  long time();
 
 }
