+++ src/test/java/com/challenge/ship/coordination/dispatchservice/MockPosition.java
@@ -2,6 +2,6 @@ package com.challenge.ship.coordination.dispatchservice;
 
 import com.challenge.ship.coordination.dispatchservice.model.TemporalPosition;
 
-public record MockPosition(int time, int x, int y) implements TemporalPosition {
+public record MockPosition(long time, int x, int y) implements TemporalPosition {
 
 }
